remove filesystem interactions from Build
